Update to openxr loader 1.0.12
