--- openxr_loader_windows/1.0.9/include/openxr/openxr_platform.h
@@ -1,401 +1,0 @@
-#ifndef OPENXR_PLATFORM_H_
-#define OPENXR_PLATFORM_H_ 1
-
-/*
-** Copyright (c) 2017-2020 The Khronos Group Inc.
-**
-** SPDX-License-Identifier: Apache-2.0
-**
-** Licensed under the Apache License, Version 2.0 (the "License");
-** you may not use this file except in compliance with the License.
-** You may obtain a copy of the License at
-**
-**     http://www.apache.org/licenses/LICENSE-2.0
-**
-** Unless required by applicable law or agreed to in writing, software
-** distributed under the License is distributed on an "AS IS" BASIS,
-** WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-** See the License for the specific language governing permissions and
-** limitations under the License.
-*/
-
-/*
-** This header is generated from the Khronos OpenXR XML API Registry.
-**
-*/
-
-#include "openxr.h"
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-
-#ifdef XR_USE_PLATFORM_ANDROID
-
-#define XR_KHR_android_thread_settings 1
-#define XR_KHR_android_thread_settings_SPEC_VERSION 5
-#define XR_KHR_ANDROID_THREAD_SETTINGS_EXTENSION_NAME "XR_KHR_android_thread_settings"
-
-typedef enum XrAndroidThreadTypeKHR {
-    XR_ANDROID_THREAD_TYPE_APPLICATION_MAIN_KHR = 1,
-    XR_ANDROID_THREAD_TYPE_APPLICATION_WORKER_KHR = 2,
-    XR_ANDROID_THREAD_TYPE_RENDERER_MAIN_KHR = 3,
-    XR_ANDROID_THREAD_TYPE_RENDERER_WORKER_KHR = 4,
-    XR_ANDROID_THREAD_TYPE_MAX_ENUM_KHR = 0x7FFFFFFF
-} XrAndroidThreadTypeKHR;
-typedef XrResult (XRAPI_PTR *PFN_xrSetAndroidApplicationThreadKHR)(XrSession session, XrAndroidThreadTypeKHR threadType, uint32_t threadId);
-
-#ifndef XR_NO_PROTOTYPES
-XRAPI_ATTR XrResult XRAPI_CALL xrSetAndroidApplicationThreadKHR(
-    XrSession                                   session,
-    XrAndroidThreadTypeKHR                      threadType,
-    uint32_t                                    threadId);
-#endif
-#endif /* XR_USE_PLATFORM_ANDROID */
-
-#ifdef XR_USE_PLATFORM_ANDROID
-
-#define XR_KHR_android_surface_swapchain 1
-#define XR_KHR_android_surface_swapchain_SPEC_VERSION 4
-#define XR_KHR_ANDROID_SURFACE_SWAPCHAIN_EXTENSION_NAME "XR_KHR_android_surface_swapchain"
-typedef XrResult (XRAPI_PTR *PFN_xrCreateSwapchainAndroidSurfaceKHR)(XrSession session, const XrSwapchainCreateInfo* info, XrSwapchain* swapchain, jobject* surface);
-
-#ifndef XR_NO_PROTOTYPES
-XRAPI_ATTR XrResult XRAPI_CALL xrCreateSwapchainAndroidSurfaceKHR(
-    XrSession                                   session,
-    const XrSwapchainCreateInfo*                info,
-    XrSwapchain*                                swapchain,
-    jobject*                                    surface);
-#endif
-#endif /* XR_USE_PLATFORM_ANDROID */
-
-#ifdef XR_USE_PLATFORM_ANDROID
-
-#define XR_KHR_android_create_instance 1
-#define XR_KHR_android_create_instance_SPEC_VERSION 3
-#define XR_KHR_ANDROID_CREATE_INSTANCE_EXTENSION_NAME "XR_KHR_android_create_instance"
-typedef struct XrInstanceCreateInfoAndroidKHR {
-    XrStructureType             type;
-    const void* XR_MAY_ALIAS    next;
-    void* XR_MAY_ALIAS          applicationVM;
-    void* XR_MAY_ALIAS          applicationActivity;
-} XrInstanceCreateInfoAndroidKHR;
-
-#endif /* XR_USE_PLATFORM_ANDROID */
-
-#ifdef XR_USE_GRAPHICS_API_VULKAN
-
-#define XR_KHR_vulkan_swapchain_format_list 1
-#define XR_KHR_vulkan_swapchain_format_list_SPEC_VERSION 2
-#define XR_KHR_VULKAN_SWAPCHAIN_FORMAT_LIST_EXTENSION_NAME "XR_KHR_vulkan_swapchain_format_list"
-typedef struct XrVulkanSwapchainFormatListCreateInfoKHR {
-    XrStructureType             type;
-    const void* XR_MAY_ALIAS    next;
-    uint32_t                    viewFormatCount;
-    const VkFormat*             viewFormats;
-} XrVulkanSwapchainFormatListCreateInfoKHR;
-
-#endif /* XR_USE_GRAPHICS_API_VULKAN */
-
-#ifdef XR_USE_GRAPHICS_API_OPENGL
-
-#define XR_KHR_opengl_enable 1
-#define XR_KHR_opengl_enable_SPEC_VERSION 8
-#define XR_KHR_OPENGL_ENABLE_EXTENSION_NAME "XR_KHR_opengl_enable"
-#ifdef XR_USE_PLATFORM_WIN32
-typedef struct XrGraphicsBindingOpenGLWin32KHR {
-    XrStructureType             type;
-    const void* XR_MAY_ALIAS    next;
-    HDC                         hDC;
-    HGLRC                       hGLRC;
-} XrGraphicsBindingOpenGLWin32KHR;
-#endif // XR_USE_PLATFORM_WIN32
-
-#ifdef XR_USE_PLATFORM_XLIB
-typedef struct XrGraphicsBindingOpenGLXlibKHR {
-    XrStructureType             type;
-    const void* XR_MAY_ALIAS    next;
-    Display*                    xDisplay;
-    uint32_t                    visualid;
-    GLXFBConfig                 glxFBConfig;
-    GLXDrawable                 glxDrawable;
-    GLXContext                  glxContext;
-} XrGraphicsBindingOpenGLXlibKHR;
-#endif // XR_USE_PLATFORM_XLIB
-
-#ifdef XR_USE_PLATFORM_XCB
-typedef struct XrGraphicsBindingOpenGLXcbKHR {
-    XrStructureType             type;
-    const void* XR_MAY_ALIAS    next;
-    xcb_connection_t*           connection;
-    uint32_t                    screenNumber;
-    xcb_glx_fbconfig_t          fbconfigid;
-    xcb_visualid_t              visualid;
-    xcb_glx_drawable_t          glxDrawable;
-    xcb_glx_context_t           glxContext;
-} XrGraphicsBindingOpenGLXcbKHR;
-#endif // XR_USE_PLATFORM_XCB
-
-#ifdef XR_USE_PLATFORM_WAYLAND
-typedef struct XrGraphicsBindingOpenGLWaylandKHR {
-    XrStructureType             type;
-    const void* XR_MAY_ALIAS    next;
-    struct wl_display*          display;
-} XrGraphicsBindingOpenGLWaylandKHR;
-#endif // XR_USE_PLATFORM_WAYLAND
-
-typedef struct XrSwapchainImageOpenGLKHR {
-    XrStructureType       type;
-    void* XR_MAY_ALIAS    next;
-    uint32_t              image;
-} XrSwapchainImageOpenGLKHR;
-
-typedef struct XrGraphicsRequirementsOpenGLKHR {
-    XrStructureType       type;
-    void* XR_MAY_ALIAS    next;
-    XrVersion             minApiVersionSupported;
-    XrVersion             maxApiVersionSupported;
-} XrGraphicsRequirementsOpenGLKHR;
-
-typedef XrResult (XRAPI_PTR *PFN_xrGetOpenGLGraphicsRequirementsKHR)(XrInstance instance, XrSystemId systemId, XrGraphicsRequirementsOpenGLKHR* graphicsRequirements);
-
-#ifndef XR_NO_PROTOTYPES
-XRAPI_ATTR XrResult XRAPI_CALL xrGetOpenGLGraphicsRequirementsKHR(
-    XrInstance                                  instance,
-    XrSystemId                                  systemId,
-    XrGraphicsRequirementsOpenGLKHR*            graphicsRequirements);
-#endif
-#endif /* XR_USE_GRAPHICS_API_OPENGL */
-
-#ifdef XR_USE_GRAPHICS_API_OPENGL_ES
-
-#define XR_KHR_opengl_es_enable 1
-#define XR_KHR_opengl_es_enable_SPEC_VERSION 6
-#define XR_KHR_OPENGL_ES_ENABLE_EXTENSION_NAME "XR_KHR_opengl_es_enable"
-#ifdef XR_USE_PLATFORM_ANDROID
-typedef struct XrGraphicsBindingOpenGLESAndroidKHR {
-    XrStructureType             type;
-    const void* XR_MAY_ALIAS    next;
-    EGLDisplay                  display;
-    EGLConfig                   config;
-    EGLContext                  context;
-} XrGraphicsBindingOpenGLESAndroidKHR;
-#endif // XR_USE_PLATFORM_ANDROID
-
-typedef struct XrSwapchainImageOpenGLESKHR {
-    XrStructureType       type;
-    void* XR_MAY_ALIAS    next;
-    uint32_t              image;
-} XrSwapchainImageOpenGLESKHR;
-
-typedef struct XrGraphicsRequirementsOpenGLESKHR {
-    XrStructureType       type;
-    void* XR_MAY_ALIAS    next;
-    XrVersion             minApiVersionSupported;
-    XrVersion             maxApiVersionSupported;
-} XrGraphicsRequirementsOpenGLESKHR;
-
-typedef XrResult (XRAPI_PTR *PFN_xrGetOpenGLESGraphicsRequirementsKHR)(XrInstance instance, XrSystemId systemId, XrGraphicsRequirementsOpenGLESKHR* graphicsRequirements);
-
-#ifndef XR_NO_PROTOTYPES
-XRAPI_ATTR XrResult XRAPI_CALL xrGetOpenGLESGraphicsRequirementsKHR(
-    XrInstance                                  instance,
-    XrSystemId                                  systemId,
-    XrGraphicsRequirementsOpenGLESKHR*          graphicsRequirements);
-#endif
-#endif /* XR_USE_GRAPHICS_API_OPENGL_ES */
-
-#ifdef XR_USE_GRAPHICS_API_VULKAN
-
-#define XR_KHR_vulkan_enable 1
-#define XR_KHR_vulkan_enable_SPEC_VERSION 6
-#define XR_KHR_VULKAN_ENABLE_EXTENSION_NAME "XR_KHR_vulkan_enable"
-typedef struct XrGraphicsBindingVulkanKHR {
-    XrStructureType             type;
-    const void* XR_MAY_ALIAS    next;
-    VkInstance                  instance;
-    VkPhysicalDevice            physicalDevice;
-    VkDevice                    device;
-    uint32_t                    queueFamilyIndex;
-    uint32_t                    queueIndex;
-} XrGraphicsBindingVulkanKHR;
-
-typedef struct XrSwapchainImageVulkanKHR {
-    XrStructureType       type;
-    void* XR_MAY_ALIAS    next;
-    VkImage               image;
-} XrSwapchainImageVulkanKHR;
-
-typedef struct XrGraphicsRequirementsVulkanKHR {
-    XrStructureType       type;
-    void* XR_MAY_ALIAS    next;
-    XrVersion             minApiVersionSupported;
-    XrVersion             maxApiVersionSupported;
-} XrGraphicsRequirementsVulkanKHR;
-
-typedef XrResult (XRAPI_PTR *PFN_xrGetVulkanInstanceExtensionsKHR)(XrInstance instance, XrSystemId systemId, uint32_t bufferCapacityInput, uint32_t* bufferCountOutput, char* buffer);
-typedef XrResult (XRAPI_PTR *PFN_xrGetVulkanDeviceExtensionsKHR)(XrInstance instance, XrSystemId systemId, uint32_t bufferCapacityInput, uint32_t* bufferCountOutput, char* buffer);
-typedef XrResult (XRAPI_PTR *PFN_xrGetVulkanGraphicsDeviceKHR)(XrInstance instance, XrSystemId systemId, VkInstance vkInstance, VkPhysicalDevice* vkPhysicalDevice);
-typedef XrResult (XRAPI_PTR *PFN_xrGetVulkanGraphicsRequirementsKHR)(XrInstance instance, XrSystemId systemId, XrGraphicsRequirementsVulkanKHR* graphicsRequirements);
-
-#ifndef XR_NO_PROTOTYPES
-XRAPI_ATTR XrResult XRAPI_CALL xrGetVulkanInstanceExtensionsKHR(
-    XrInstance                                  instance,
-    XrSystemId                                  systemId,
-    uint32_t                                    bufferCapacityInput,
-    uint32_t*                                   bufferCountOutput,
-    char*                                       buffer);
-
-XRAPI_ATTR XrResult XRAPI_CALL xrGetVulkanDeviceExtensionsKHR(
-    XrInstance                                  instance,
-    XrSystemId                                  systemId,
-    uint32_t                                    bufferCapacityInput,
-    uint32_t*                                   bufferCountOutput,
-    char*                                       buffer);
-
-XRAPI_ATTR XrResult XRAPI_CALL xrGetVulkanGraphicsDeviceKHR(
-    XrInstance                                  instance,
-    XrSystemId                                  systemId,
-    VkInstance                                  vkInstance,
-    VkPhysicalDevice*                           vkPhysicalDevice);
-
-XRAPI_ATTR XrResult XRAPI_CALL xrGetVulkanGraphicsRequirementsKHR(
-    XrInstance                                  instance,
-    XrSystemId                                  systemId,
-    XrGraphicsRequirementsVulkanKHR*            graphicsRequirements);
-#endif
-#endif /* XR_USE_GRAPHICS_API_VULKAN */
-
-#ifdef XR_USE_GRAPHICS_API_D3D11
-
-#define XR_KHR_D3D11_enable 1
-#define XR_KHR_D3D11_enable_SPEC_VERSION  4
-#define XR_KHR_D3D11_ENABLE_EXTENSION_NAME "XR_KHR_D3D11_enable"
-typedef struct XrGraphicsBindingD3D11KHR {
-    XrStructureType             type;
-    const void* XR_MAY_ALIAS    next;
-    ID3D11Device*               device;
-} XrGraphicsBindingD3D11KHR;
-
-typedef struct XrSwapchainImageD3D11KHR {
-     XrStructureType      type;
-    void* XR_MAY_ALIAS    next;
-    ID3D11Texture2D*      texture;
-} XrSwapchainImageD3D11KHR;
-
-typedef struct XrGraphicsRequirementsD3D11KHR {
-    XrStructureType       type;
-    void* XR_MAY_ALIAS    next;
-    LUID                  adapterLuid;
-    D3D_FEATURE_LEVEL     minFeatureLevel;
-} XrGraphicsRequirementsD3D11KHR;
-
-typedef XrResult (XRAPI_PTR *PFN_xrGetD3D11GraphicsRequirementsKHR)(XrInstance instance, XrSystemId systemId, XrGraphicsRequirementsD3D11KHR* graphicsRequirements);
-
-#ifndef XR_NO_PROTOTYPES
-XRAPI_ATTR XrResult XRAPI_CALL xrGetD3D11GraphicsRequirementsKHR(
-    XrInstance                                  instance,
-    XrSystemId                                  systemId,
-    XrGraphicsRequirementsD3D11KHR*             graphicsRequirements);
-#endif
-#endif /* XR_USE_GRAPHICS_API_D3D11 */
-
-#ifdef XR_USE_GRAPHICS_API_D3D12
-
-#define XR_KHR_D3D12_enable 1
-#define XR_KHR_D3D12_enable_SPEC_VERSION  6
-#define XR_KHR_D3D12_ENABLE_EXTENSION_NAME "XR_KHR_D3D12_enable"
-typedef struct XrGraphicsBindingD3D12KHR {
-    XrStructureType             type;
-    const void* XR_MAY_ALIAS    next;
-    ID3D12Device*               device;
-    ID3D12CommandQueue*         queue;
-} XrGraphicsBindingD3D12KHR;
-
-typedef struct XrSwapchainImageD3D12KHR {
-     XrStructureType      type;
-    void* XR_MAY_ALIAS    next;
-    ID3D12Resource*       texture;
-} XrSwapchainImageD3D12KHR;
-
-typedef struct XrGraphicsRequirementsD3D12KHR {
-    XrStructureType       type;
-    void* XR_MAY_ALIAS    next;
-    LUID                  adapterLuid;
-    D3D_FEATURE_LEVEL     minFeatureLevel;
-} XrGraphicsRequirementsD3D12KHR;
-
-typedef XrResult (XRAPI_PTR *PFN_xrGetD3D12GraphicsRequirementsKHR)(XrInstance instance, XrSystemId systemId, XrGraphicsRequirementsD3D12KHR* graphicsRequirements);
-
-#ifndef XR_NO_PROTOTYPES
-XRAPI_ATTR XrResult XRAPI_CALL xrGetD3D12GraphicsRequirementsKHR(
-    XrInstance                                  instance,
-    XrSystemId                                  systemId,
-    XrGraphicsRequirementsD3D12KHR*             graphicsRequirements);
-#endif
-#endif /* XR_USE_GRAPHICS_API_D3D12 */
-
-#ifdef XR_USE_PLATFORM_WIN32
-
-#define XR_KHR_win32_convert_performance_counter_time 1
-#define XR_KHR_win32_convert_performance_counter_time_SPEC_VERSION 1
-#define XR_KHR_WIN32_CONVERT_PERFORMANCE_COUNTER_TIME_EXTENSION_NAME "XR_KHR_win32_convert_performance_counter_time"
-typedef XrResult (XRAPI_PTR *PFN_xrConvertWin32PerformanceCounterToTimeKHR)(XrInstance instance, const LARGE_INTEGER* performanceCounter, XrTime* time);
-typedef XrResult (XRAPI_PTR *PFN_xrConvertTimeToWin32PerformanceCounterKHR)(XrInstance instance, XrTime   time, LARGE_INTEGER* performanceCounter);
-
-#ifndef XR_NO_PROTOTYPES
-XRAPI_ATTR XrResult XRAPI_CALL xrConvertWin32PerformanceCounterToTimeKHR(
-    XrInstance                                  instance,
-    const LARGE_INTEGER*                        performanceCounter,
-    XrTime*                                     time);
-
-XRAPI_ATTR XrResult XRAPI_CALL xrConvertTimeToWin32PerformanceCounterKHR(
-    XrInstance                                  instance,
-    XrTime                                      time,
-    LARGE_INTEGER*                              performanceCounter);
-#endif
-#endif /* XR_USE_PLATFORM_WIN32 */
-
-#ifdef XR_USE_TIMESPEC
-
-#define XR_KHR_convert_timespec_time 1
-#define XR_KHR_convert_timespec_time_SPEC_VERSION 1
-#define XR_KHR_CONVERT_TIMESPEC_TIME_EXTENSION_NAME "XR_KHR_convert_timespec_time"
-typedef XrResult (XRAPI_PTR *PFN_xrConvertTimespecTimeToTimeKHR)(XrInstance instance, const struct timespec* timespecTime, XrTime* time);
-typedef XrResult (XRAPI_PTR *PFN_xrConvertTimeToTimespecTimeKHR)(XrInstance instance, XrTime   time, struct timespec* timespecTime);
-
-#ifndef XR_NO_PROTOTYPES
-XRAPI_ATTR XrResult XRAPI_CALL xrConvertTimespecTimeToTimeKHR(
-    XrInstance                                  instance,
-    const struct timespec*                      timespecTime,
-    XrTime*                                     time);
-
-XRAPI_ATTR XrResult XRAPI_CALL xrConvertTimeToTimespecTimeKHR(
-    XrInstance                                  instance,
-    XrTime                                      time,
-    struct timespec*                            timespecTime);
-#endif
-#endif /* XR_USE_TIMESPEC */
-
-#ifdef XR_USE_PLATFORM_EGL
-
-#define XR_MNDX_egl_enable 1
-#define XR_MNDX_egl_enable_SPEC_VERSION   1
-#define XR_MNDX_EGL_ENABLE_EXTENSION_NAME "XR_MNDX_egl_enable"
-typedef struct XrGraphicsBindingEGLMNDX {
-    XrStructureType             type;
-    const void* XR_MAY_ALIAS    next;
-    PFNEGLGETPROCADDRESSPROC    getProcAddress;
-    EGLDisplay                  display;
-    EGLConfig                   config;
-    EGLContext                  context;
-} XrGraphicsBindingEGLMNDX;
-
-#endif /* XR_USE_PLATFORM_EGL */
-
-#ifdef __cplusplus
-}
-#endif
-
-#endif
